去掉Control.InvokeRequired的判断会导致线程不安全的问题。具体原因如下：
Access to Windows Forms controls is not inherently thread safe. If you have 2 or more threads manipulating the state of a control, it is possible to force the control into
an inconsistent state. Other thread-related bugs are possible, such as race conditions and deadlocks.

It is unsafe to call a control from a thread, other tha the one that create the control, without using the "Invoke" method.
The .NET framework helps you detect when you are accessing your controls in a manner that is not thread safe. When you are running your application in the debugger, and a thread, other than the oen which created a control,
tries to access that control, the debugger raises an InvalidOperationException.
This exception occurs reliably during debugging and, under some circumstances, at run time.